# packages/server-go/docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: DevBox SDK Server API
  description: |
    A comprehensive API for managing processes, sessions, files, and providing real-time monitoring capabilities.

    The DevBox SDK Server provides HTTP endpoints for:
    - **File Operations**: Read, write, delete, and list files with security constraints
    - **Process Management**: Execute processes synchronously or asynchronously with log monitoring
    - **Session Management**: Create and manage interactive shell sessions
    - **Real-time Communication**: WebSocket connections for live log streaming
    - **Health Monitoring**: Health check and readiness endpoints
    - **Port Monitoring**: Monitor listening ports on the system

    ## Authentication
    All API endpoints (except health checks) require Bearer token authentication:

    ```http
    Authorization: Bearer <your-token>
    ```

    ## Error Handling
    The API uses standard HTTP status codes and returns consistent error responses:

    ```json
    {
      "error": "Error description",
      "code": "ERROR_CODE",
      "timestamp": 1640995200000
    }
    ```

  version: 1.0.0
  contact:
    name: DevBox SDK Team
    url: https://github.com/labring/devbox-sdk
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:9757
    description: Development server (default port, configurable via ADDR env or -addr flag)
  - url: https://your-server.example.com
...
paths:
...
  /api/v1/process/exec:
    post:
      tags:
        - Processes
      summary: Execute process asynchronously
      description: Execute a new process asynchronously and return immediately with process ID
      security:
        - bearerAuth: []
      operationId: execProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessExecRequest'
            example:
              command: "ls"
              args: ["-la", "/tmp"]
              cwd: "/home/user"
              env:
                PATH: "/usr/bin:/bin"
                DEBUG: "true"
              timeout: 300
      responses:
        '200':
          description: Process started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessExecResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Failed to start process
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/process/sync-stream:
    post:
      tags:
        - Processes
      summary: Execute process with streaming
      description: Execute a process synchronously with Server-Sent Events streaming for real-time output
      security:
        - bearerAuth: []
      operationId: execProcessSyncStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncExecutionRequest'
      responses:
        '200':
          description: Process streaming started
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream with process output
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/sessions/{id}/exec:
    post:
      tags:
        - Sessions
      summary: Execute command in session
      description: Execute a command in an active session
      security:
        - bearerAuth: []
      operationId: sessionExec
      parameters:
        - name: id
          in: path
          description: Session ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionExecRequest'
            example:
              command: "ls -la"
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionExecResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Response:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        error:
          type: string
          description: Error message (if any)
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error description
        code:
          type: string
          description: Error code
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
      required:
        - error
        - timestamp
...
    ProcessExecRequest:
      type: object
      properties:
        command:
          type: string
          description: Command to execute
          example: "ls"
        args:
          type: array
          items:
            type: string
          description: Command arguments
          example: ["-la", "/tmp"]
        cwd:
          type: string
          description: Working directory
          example: "/home/user"
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
          example:
            PATH: "/usr/bin:/bin"
            DEBUG: "true"
        shell:
          type: string
          description: Shell to use for execution
          example: "/bin/bash"
        timeout:
          type: integer
          description: Timeout in seconds
          example: 300
      required:
        - command
    ProcessExecResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            processId:
              type: string
              description: Generated process ID
              example: "550e8400-e29b-41d4-a716-446655440000"
            pid:
              type: integer
              description: System process ID
              example: 12345
            status:
              type: string
              description: Process status
              example: "running"
            exitCode:
              type: integer
              description: Exit code (if completed)
              example: 0
      required:
        - success
        - processId
        - pid
        - status
    SyncExecutionRequest:
      type: object
      properties:
        command:
          type: string
          description: Command to execute
          example: "echo"
        args:
          type: array
          items:
            type: string
          description: Command arguments
          example: ["Hello World"]
        cwd:
          type: string
          description: Working directory
          example: "/home/user"
        env:
          type: object
          additionalProperties:
            type: string
          description: Environment variables
          example:
            PATH: "/usr/bin:/bin"
        shell:
          type: string
          description: Shell to use for execution
          example: "/bin/bash"
        timeout:
          type: integer
          description: Timeout in seconds
          example: 30
      required:
        - command
...
    SessionExecRequest:
      type: object
      properties:
        command:
          type: string
          description: Command to execute in session
          example: "ls -la"
      required:
        - command
    SessionExecResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            exitCode:
              type: integer
              description: Command exit code
              example: 0
            stdout:
              type: string
              description: Command output (standard output)
              example: ""
            stderr:
              type: string
              description: Error output (standard error)
              example: ""
            duration:
              type: integer
              format: int64
              description: Execution duration in milliseconds
              example: 0
      required:
        - success
        - exitCode
        - stdout
        - stderr
        - duration
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid request parameters"
            code: "INVALID_REQUEST"
            timestamp: 1640995200000
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
            timestamp: 1640995200000
...